service2_subscriber/requirements.txt
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0
